list组件滚动偏移量float改成double，提高上报开发者的偏移量精度

+++ frameworks/core/components_ng/pattern/scroll/scroll_pattern.h
@@ -105,7 +105,7 @@ public:
         return currentOffset_;
     }
 
-    float GetTotalOffset() const override
+    double GetTotalOffset() const override
     {
         return -currentOffset_;
     }
